# gcs: Implement version 2 filters.